% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFinalFit.R
\name{computeFinalFit}
\alias{computeFinalFit}
\title{FSCRE Final Robust Fitting (Internal)}
\usage{
computeFinalFit(x.imp, y.imp, active.sets, compute_coef)
}
\arguments{
\item{x.imp}{Imputed design matrix.}

\item{y.imp}{Imputed response vector.}

\item{active.sets}{List of integer vectors (indices of selected variables).}

\item{compute_coef}{Logical.}
}
\value{
A list containing `coefficients` (list of vectors) and `intercepts` (vector).
}
\description{
Implements Stage 3: Fitting robust MM-estimators to the selected variable sets.
}
\keyword{internal}
