% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{checkInputData}
\alias{checkInputData}
\title{Check Input Data for srlars Function}
\usage{
checkInputData(x, y, n_models, tolerance, max_predictors, robust, compute_coef)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Response vector.}

\item{n_models}{Number of models in the ensemble.}

\item{tolerance}{Relative improvement tolerance for stopping.}

\item{max_predictors}{Maximum total number of variables to select.}

\item{robust}{Logical.}

\item{compute_coef}{Logical.}
}
\value{
NULL. Stops execution with an error message if invalid inputs are detected.
}
\description{
Internal helper function to validate arguments passed to \code{srlars}.
Checks types, dimensions, and logical constraints.
}
\keyword{internal}
